requirements.txt
asgiref==3.8.1
certifi==2024.8.30
charset-normalizer==3.3.2
Django==5.1.1
idna==3.10
pillow==10.4.0
plyer==2.1.0
psycopg2-binary==2.9.9
pymemcache==4.0.0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
sqlparse==0.5.1
tzdata==2024.2
urllib3==2.2.3
